Address PR feedback on auto-translation experiment
- Capture the banner's language_changed event inline instead of extending
  SetOriginalLanguage's signature
- Read the assignment cookie inside CSPostHogProvider (cookie is no longer
  httpOnly) so layout.tsx stays untouched; parse/serialize helpers move to
  the client-safe constants module
- Fold the posthog-node singleton into posthog.server.ts, drop the unused
  cookie-based getFeatureFlag and the extra posthog_node_client.ts file
