+++ front_end/src/app/(main)/components/content_translated_banner/index.tsx
@@ -2,6 +2,7 @@
 
 import { usePathname, useRouter } from "next/navigation";
 import { useLocale, useTranslations } from "next-intl";
+import posthog from "posthog-js";
 import { FC, useEffect } from "react";
 
 import { SetOriginalLanguage as setOriginalLanguage } from "@/components/language_menu";
@@ -46,9 +47,19 @@ const ContentTranslatedBanner: FC<{
             className="inline whitespace-nowrap"
             variant="link"
             size="xs"
-            onClick={() =>
-              setOriginalLanguage(params, router, pathname, locale)
-            }
+            onClick={() => {
+              // Beacon transport so the event survives the hard navigation
+              posthog.capture(
+                "language_changed",
+                {
+                  previous_language: locale,
+                  new_language: "original",
+                  source: "content_translated_banner",
+                },
+                { transport: "sendBeacon" }
+              );
+              setOriginalLanguage(params, router, pathname);
+            }}
           >
             {t("showOriginalContent")}
           </Button>
