+++ front_end/src/contexts/posthog_context.tsx
@@ -7,24 +7,44 @@ import { getAnalyticsCookieConsentGiven } from "@/app/(main)/components/cookies_
 import SuspendedPostHogPageView from "@/components/posthog_page_view";
 import { getPublicSetting } from "@/components/public_settings_script";
 import {
+  AUTOTRANSLATION_COOKIE_NAME,
   AUTOTRANSLATION_FLAG_KEY,
   AutotranslationAssignment,
+  parseAssignment,
 } from "@/constants/experiments";
+
+// The auto-translation experiment assignment is pinned in a first-party
+// cookie by the middleware (proxy.ts) when an eligible visitor is enrolled
+function getAutotranslationAssignment(): AutotranslationAssignment | null {
+  const raw = document.cookie
+    .split("; ")
+    .find((cookie) => cookie.startsWith(`${AUTOTRANSLATION_COOKIE_NAME}=`))
+    ?.slice(AUTOTRANSLATION_COOKIE_NAME.length + 1);
+  if (!raw) return null;
+
+  try {
+    // Next.js URL-encodes cookie values when setting; document.cookie
+    // returns them still encoded
+    return parseAssignment(decodeURIComponent(raw));
+  } catch {
+    return null;
+  }
+}
 
 function CSPostHogProvider({
   children,
   locale,
-  autotranslationAssignment,
 }: {
   children: ReactNode;
   locale: string;
-  autotranslationAssignment?: AutotranslationAssignment;
 }) {
   useEffect(() => {
     const PUBLIC_POSTHOG_KEY = getPublicSetting("PUBLIC_POSTHOG_KEY");
     const PUBLIC_POSTHOG_BASE_URL = getPublicSetting("PUBLIC_POSTHOG_BASE_URL");
 
     if (PUBLIC_POSTHOG_KEY) {
+      const autotranslationAssignment = getAutotranslationAssignment();
+
       posthog.init(PUBLIC_POSTHOG_KEY, {
         api_host: PUBLIC_POSTHOG_BASE_URL,
         ui_host: "https://us.posthog.com",
@@ -55,7 +75,6 @@ function CSPostHogProvider({
         posthog.getFeatureFlag(AUTOTRANSLATION_FLAG_KEY);
       }
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
