+++ front_end/src/utils/posthog.server.ts
@@ -1,31 +1,42 @@
 import "server-only";
 
-import { cookies } from "next/headers";
+import { PostHog } from "posthog-node";
 
-import { getPostHogServerClient } from "./posthog_node_client";
 import { getPublicSettings } from "./public_settings.server";
 
-export async function getFeatureFlag(
-  flagName: string,
-  defaultValue: boolean | string = false
-): Promise<boolean | string> {
+let posthogClient: PostHog | null = null;
+
+// Kept free of next/headers so it is importable from proxy.ts (middleware)
+function getPostHogServerClient(): PostHog | null {
+  const { PUBLIC_POSTHOG_KEY, PUBLIC_POSTHOG_BASE_URL } = getPublicSettings();
+  if (!PUBLIC_POSTHOG_KEY) return null;
+
+  if (!posthogClient) {
+    posthogClient = new PostHog(PUBLIC_POSTHOG_KEY, {
+      host: PUBLIC_POSTHOG_BASE_URL || "https://us.i.posthog.com",
+      flushAt: 1,
+      flushInterval: 0,
+      featureFlagsRequestTimeoutMs: 1500,
+    });
+  }
+  return posthogClient;
+}
+
+export async function getFeatureFlagVariantForDistinctId(
+  flagKey: string,
+  distinctId: string
+): Promise<string | null> {
   const client = getPostHogServerClient();
-  if (!client) return defaultValue;
-
-  const { PUBLIC_POSTHOG_KEY } = getPublicSettings();
-
-  const cookieStore = await cookies();
-  const cookieName = "ph_" + PUBLIC_POSTHOG_KEY + "_posthog";
-  const cookieValue = cookieStore.get(cookieName)?.value;
-  const distinctId = cookieValue
-    ? JSON.parse(cookieValue).distinct_id
-    : "anonymous";
+  if (!client) return null;
 
   try {
-    const flag = await client.getFeatureFlag(flagName, distinctId);
-
-    return flag !== undefined ? flag : defaultValue;
+    // Exposure ($feature_flag_called) is captured client-side with the same
+    // distinct_id, so suppress the server-side event here
+    const variant = await client.getFeatureFlag(flagKey, distinctId, {
+      sendFeatureFlagEvents: false,
+    });
+    return typeof variant === "string" ? variant : null;
   } catch {
-    return defaultValue;
+    return null;
   }
 }
